# Conflicts:
#	src/RiakClient/Commands/CRDT/DataTypeResponse{TValue}.cs
#	src/RiakClient/Commands/CRDT/FetchCommandBuilder{TBuilder,TCommand,TOptions,TResponse}.cs
#	src/RiakClient/Commands/CRDT/FetchCommand{TResponse}.cs
#	src/RiakClient/Commands/CRDT/UpdateCommandBuilder{TBuilder,TCommand,TOptions,TResponse}.cs
#	src/RiakClient/Commands/CRDT/UpdateCommandOptions.cs
#	src/RiakClient/Commands/CRDT/UpdateCounter.cs
#	src/RiakClient/Commands/CRDT/UpdateCounterOptions.cs
#	src/RiakClient/Commands/CRDT/UpdateMap.cs
#	src/RiakClient/Commands/CRDT/UpdateSet.cs
#	src/RiakClient/Commands/CommandBuilder{TBuilder,TCommand,TOptions}.cs
#	src/RiakClient/Commands/CommandOptions.cs
#	src/RiakClient/Commands/Command{TOptions,TResponse}.cs
#	src/RiakClient/Commands/Command{TResponse}.cs
#	src/RiakClient/Commands/IRiakCommand.cs
#	src/RiakClient/Commands/KV/FetchPreflist.cs
#	src/RiakClient/Commands/KV/PreflistResponse.cs
#	src/RiakClient/Commands/Response.cs
#	src/RiakClient/Commands/Response{TValue}.cs
#	src/RiakClient/Commands/ServerInfoResponse.cs
#	src/RiakClient/Comms/RiakPbcSocket.cs
#	src/RiakClient/Core/SecurityManager.cs
#	src/RiakClient/Messages/RpbClasses.cs
#	src/RiakClient/Messages/riak_dt.cs
#	src/RiakClient/Messages/riak_ts.cs
#	src/RiakClient/Models/RiakDeleteOptions.cs
#	src/RiakClient/Models/RiakOptions.cs
#	src/RiakClient/RiakClient.cs
#	src/RiakClientTests.Live/App.config
#	src/RiakClientTests.Live/DataTypes/BasicMapFlagDtTests.cs
#	src/RiakClientTests.Live/RiakConfigurationTests.cs
#	src/Test/App.config
#	src/Test/Traits.cs
#	src/Test/packages.config

+++ src/Test/Integration/TestBase.cs
@@ -28,12 +28,12 @@ namespace Test.Integration
             RiakClient.DisableListBucketsWarning = true;
         }
 
-        public TestBase()
+        public TestBase(bool auth = true)
         {
 #if NOAUTH
             cluster = RiakCluster.FromConfig("riak1NodeNoAuthConfiguration");
 #else
-            if (MonoUtil.IsRunningOnMono)
+            if (auth == false || MonoUtil.IsRunningOnMono)
             {
                 cluster = RiakCluster.FromConfig("riak1NodeNoAuthConfiguration");
             }
@@ -41,32 +41,36 @@ namespace Test.Integration
             {
                 cluster = RiakCluster.FromConfig("riak1NodeConfiguration");
             }
+
 #endif
             client = cluster.CreateClient();
         }
 
-        protected abstract RiakString BucketType
+        protected virtual RiakString BucketType
         {
-            get;
+            get { return "bucketType_unset"; }
         }
 
-        protected abstract RiakString Bucket
+        protected virtual RiakString Bucket
         {
-            get;
+            get { return "bucket_unset"; }
         }
 
         [TestFixtureSetUp]
-        protected virtual void TestFixtureSetUp()
+        public virtual void TestFixtureSetUp()
         {
         }
 
         [TestFixtureTearDown]
-        protected virtual void TestFixtureTearDown()
+        public virtual void TestFixtureTearDown()
         {
-            foreach (string key in Keys)
+            if (EnumerableUtil.NotNullOrEmpty(Keys))
             {
-                var id = new RiakObjectId(BucketType, Bucket, key);
-                client.Delete(id);
+                foreach (string key in Keys)
+                {
+                    var id = new RiakObjectId(BucketType, Bucket, key);
+                    client.Delete(id);
+                }
             }
         }
